Refactor models.py into smaller files
