+++ src/meshapi/models/devices/__init__.py
@@ -1,0 +1,2 @@
+from .device import *
+from .sector import *
